rename docurator  file to docurator-object
